BACKEND/domains/user_security/password_recovery/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
httpx==0.28.1
email-validator==2.1.1  